# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/manage/consumers/list": {
      "get": {
        "operationId": "ConsumerController_list",
        "parameters": [
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "手机号码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "用户名",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取客户列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取客户列表",
        "tags": [
          "客户管理"
        ]
      }
    },
    "/api/v1/manage/consumers/page": {
      "get": {
        "operationId": "ConsumerController_page",
        "parameters": [
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "手机号码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "用户名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取客户分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取客户列表",
        "tags": [
          "客户管理"
        ]
      }
    },
    "/api/v1/manage/consumers/detail/{id}": {
      "get": {
        "operationId": "ConsumerController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取客户详情"
          },
          "404": {
            "description": "客户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取客户详情",
        "tags": [
          "客户管理"
        ]
      }
    },
    "/api/v1/manage/consumers/create": {
      "post": {
        "operationId": "ConsumerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "客户创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建客户",
        "tags": [
          "客户管理"
        ]
      }
    },
    "/api/v1/manage/consumers/update": {
      "post": {
        "operationId": "ConsumerController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsumerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "客户更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "客户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新客户",
        "tags": [
          "客户管理"
        ]
      }
    },
    "/api/v1/manage/consumers/delete": {
      "post": {
        "operationId": "ConsumerController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "客户删除成功"
          },
          "400": {
            "description": "删除客户失败"
          },
          "404": {
            "description": "客户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除客户",
        "tags": [
          "客户管理"
        ]
      }
    },
...
    "/api/v1/manage/orders/list": {
      "get": {
        "operationId": "OrderController_list",
        "parameters": [
          {
            "name": "orderNumber",
            "required": false,
            "in": "query",
            "description": "订单编号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "consumerId",
            "required": false,
            "in": "query",
            "description": "客户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": false,
            "in": "query",
            "description": "收货地址ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "订单状态",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "SHIPPED",
                "DELIVERED",
                "CANCELLED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "paymentMethod",
            "required": false,
            "in": "query",
            "description": "支付方式",
            "schema": {
              "type": "string",
              "enum": [
                "WECHAT",
                "ALIPAY",
                "CREDIT_CARD"
              ]
            }
          },
          {
            "name": "commodityId",
            "required": false,
            "in": "query",
            "description": "商品ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取订单列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取订单列表",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/orders/page": {
      "get": {
        "operationId": "OrderController_page",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取订单分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取订单列表",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/orders/detail/{id}": {
      "get": {
        "operationId": "OrderController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取订单详情"
          },
          "404": {
            "description": "订单不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取订单详情",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/orders/create": {
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建订单",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/orders/update": {
      "post": {
        "operationId": "OrderController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "订单不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新订单",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/orders/delete": {
      "post": {
        "operationId": "OrderController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单删除成功"
          },
          "400": {
            "description": "删除订单失败"
          },
          "404": {
            "description": "订单不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除订单",
        "tags": [
          "订单管理"
        ]
      }
    },
    "/api/v1/manage/order-items/list": {
      "get": {
        "operationId": "OrderItemController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取订单项列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取订单项列表",
        "tags": [
          "订单项管理"
        ]
      }
    },
    "/api/v1/manage/order-items/page": {
      "get": {
        "operationId": "OrderItemController_page",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取订单项分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取订单项列表",
        "tags": [
          "订单项管理"
        ]
      }
    },
    "/api/v1/manage/order-items/detail/{id}": {
      "get": {
        "operationId": "OrderItemController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取订单项详情"
          },
          "404": {
            "description": "订单项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取订单项详情",
        "tags": [
          "订单项管理"
        ]
      }
    },
    "/api/v1/manage/order-items/create": {
      "post": {
        "operationId": "OrderItemController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单项创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建订单项",
        "tags": [
          "订单项管理"
        ]
      }
    },
    "/api/v1/manage/order-items/update": {
      "post": {
        "operationId": "OrderItemController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单项更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "订单项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新订单项",
        "tags": [
          "订单项管理"
        ]
      }
    },
    "/api/v1/manage/order-items/delete": {
      "post": {
        "operationId": "OrderItemController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "订单项删除成功"
          },
          "400": {
            "description": "删除订单项失败"
          },
          "404": {
            "description": "订单项不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除订单项",
        "tags": [
          "订单项管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-Commerce API",
    "description": "API documentation for the e-commerce management system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateConsumerDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "手机号码"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "frontCard": {
            "type": "string",
            "description": "身份证正面照片 URL"
          },
          "backCard": {
            "type": "string",
            "description": "身份证反面照片 URL"
          },
          "userCard": {
            "type": "string",
            "description": "手持身份证照片 URL"
          }
        },
        "required": [
          "phone"
        ]
      },
      "UpdateConsumerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "客户ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "i": {
        "type": "object",
        "properties": {}
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string",
            "description": "订单编号"
          },
          "consumerId": {
            "type": "string",
            "description": "客户ID"
          },
          "addressId": {
            "type": "string",
            "description": "收货地址ID"
          },
          "totalPrice": {
            "description": "订单总价",
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "status": {
            "type": "string",
            "description": "订单状态",
            "enum": [
              "PENDING",
              "PAID",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "paymentMethod": {
            "type": "string",
            "description": "支付方式",
            "enum": [
              "WECHAT",
              "ALIPAY",
              "CREDIT_CARD"
            ]
          }
        },
        "required": [
          "orderNumber",
          "consumerId",
          "addressId",
          "totalPrice",
          "status",
          "paymentMethod"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "订单ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "订单ID"
          },
          "commodityId": {
            "type": "string",
            "description": "商品ID"
          },
          "quantity": {
            "type": "number",
            "description": "商品数量"
          },
          "price": {
            "type": "object",
            "description": "商品价格"
          }
        },
        "required": [
          "orderId",
          "commodityId",
          "quantity",
          "price"
        ]
      },
      "UpdateOrderItemDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}